refactor(extension): split lua manager
